+++ cli/run/waitForInput.ts
@@ -1,5 +1,5 @@
-import { PluginContext } from 'rollup';
-import { NormalizedInputOptions, Plugin } from '../../src/rollup/types';
+import type { PluginContext } from 'rollup';
+import type { NormalizedInputOptions, Plugin } from '../../src/rollup/types';
 import { bold } from '../../src/utils/colors';
 import { stderr } from '../logging';
 
@@ -8,9 +8,9 @@ export function waitForInputPlugin(): Plugin {
 		async buildStart(this: PluginContext, options: NormalizedInputOptions) {
 			const inputSpecifiers = Array.isArray(options.input)
 				? options.input
-				: Object.keys(options.input as { [entryAlias: string]: string });
+				: Object.keys(options.input);
 
-			let lastAwaitedSpecifier = null;
+			let lastAwaitedSpecifier: string | null = null;
 			checkSpecifiers: while (true) {
 				for (const specifier of inputSpecifiers) {
 					if ((await this.resolve(specifier)) === null) {