+++ src/utils/deconflictChunk.ts
@@ -1,11 +1,11 @@
-import Chunk from '../Chunk';
+import type Chunk from '../Chunk';
 import ExternalModule from '../ExternalModule';
-import Module from '../Module';
-import ChildScope from '../ast/scopes/ChildScope';
+import type Module from '../Module';
+import type ChildScope from '../ast/scopes/ChildScope';
 import ExportDefaultVariable from '../ast/variables/ExportDefaultVariable';
-import SyntheticNamedExportVariable from '../ast/variables/SyntheticNamedExportVariable';
-import Variable from '../ast/variables/Variable';
-import { GetInterop, InternalModuleFormat } from '../rollup/types';
+import type SyntheticNamedExportVariable from '../ast/variables/SyntheticNamedExportVariable';
+import type Variable from '../ast/variables/Variable';
+import type { GetInterop, InternalModuleFormat } from '../rollup/types';
 import {
 	canDefaultBeTakenFromNamespace,
 	defaultInteropHelpersByInteropType,
@@ -15,21 +15,21 @@ import {
 import { getSafeName } from './safeName';
 
 export interface DependenciesToBeDeconflicted {
-	deconflictedDefault: Set<ExternalModule>;
-	deconflictedNamespace: Set<ExternalModule | Chunk>;
-	dependencies: Set<ExternalModule | Chunk>;
+	deconflictedDefault: ReadonlySet<ExternalModule>;
+	deconflictedNamespace: ReadonlySet<ExternalModule | Chunk>;
+	dependencies: ReadonlySet<ExternalModule | Chunk>;
 }
 
 const DECONFLICT_IMPORTED_VARIABLES_BY_FORMAT: {
 	[format in InternalModuleFormat]: (
 		usedNames: Set<string>,
-		imports: Set<Variable>,
+		imports: ReadonlySet<Variable>,
 		dependenciesToBeDeconflicted: DependenciesToBeDeconflicted,
 		interop: GetInterop,
 		preserveModules: boolean,
 		externalLiveBindings: boolean,
-		chunkByModule: Map<Module, Chunk>,
-		syntheticExports: Set<SyntheticNamedExportVariable>
+		chunkByModule: ReadonlyMap<Module, Chunk>,
+		syntheticExports: ReadonlySet<SyntheticNamedExportVariable>
 	) => void;
 } = {
 	amd: deconflictImportsOther,
@@ -43,16 +43,16 @@ const DECONFLICT_IMPORTED_VARIABLES_BY_FORMAT: {
 export function deconflictChunk(
 	modules: readonly Module[],
 	dependenciesToBeDeconflicted: DependenciesToBeDeconflicted,
-	imports: Set<Variable>,
+	imports: ReadonlySet<Variable>,
 	usedNames: Set<string>,
 	format: InternalModuleFormat,
 	interop: GetInterop,
 	preserveModules: boolean,
 	externalLiveBindings: boolean,
-	chunkByModule: Map<Module, Chunk>,
-	syntheticExports: Set<SyntheticNamedExportVariable>,
-	exportNamesByVariable: ReadonlyMap<Variable, string[]>,
-	accessedGlobalsByScope: ReadonlyMap<ChildScope, Set<string>>,
+	chunkByModule: ReadonlyMap<Module, Chunk>,
+	syntheticExports: ReadonlySet<SyntheticNamedExportVariable>,
+	exportNamesByVariable: ReadonlyMap<Variable, readonly string[]>,
+	accessedGlobalsByScope: ReadonlyMap<ChildScope, ReadonlySet<string>>,
 	includedNamespaces: ReadonlySet<Module>
 ): void {
 	const reversedModules = modules.slice().reverse();
@@ -88,8 +88,8 @@ function deconflictImportsEsmOrSystem(
 	_interop: GetInterop,
 	preserveModules: boolean,
 	_externalLiveBindings: boolean,
-	chunkByModule: Map<Module, Chunk>,
-	syntheticExports: Set<SyntheticNamedExportVariable>
+	chunkByModule: ReadonlyMap<Module, Chunk>,
+	syntheticExports: ReadonlySet<SyntheticNamedExportVariable>
 ) {
 	// This is needed for namespace reexports
 	for (const dependency of dependenciesToBeDeconflicted.dependencies) {
@@ -128,12 +128,12 @@ function deconflictImportsEsmOrSystem(
 
 function deconflictImportsOther(
 	usedNames: Set<string>,
-	imports: Set<Variable>,
+	imports: ReadonlySet<Variable>,
 	{ deconflictedDefault, deconflictedNamespace, dependencies }: DependenciesToBeDeconflicted,
 	interop: GetInterop,
 	preserveModules: boolean,
 	externalLiveBindings: boolean,
-	chunkByModule: Map<Module, Chunk>
+	chunkByModule: ReadonlyMap<Module, Chunk>
 ): void {
 	for (const chunkOrExternalModule of dependencies) {
 		chunkOrExternalModule.variableName = getSafeName(