+++ test/form/samples/merge-namespaces/reexport.js
@@ -1,3 +1,4 @@
 export * from 'external1';
 export * from './synthetic';
 export * from 'external2';
+export const __synthetic = { module: 'reexport' };